Moved workflow task order modal to a separate component to fix a bug in the update
